Share more of job testing across OSs
